pytorch_points/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
anaconda=custom=py37_0
cuda92=1.0=0
faiss-gpu=1.5.2
imageio=2.5.0=py37_0
jsonschema=3.0.1=py37_0
matplotlib=3.0.2
matplotlib-base=3.0.2
mkl=2019.3=199
mkl_fft=1.0.12=py37h516909a_1
mkl_random=1.0.4=py37hf2d7682_0
nccl=1.3.5=cuda10.0_0
numpy=1.16.4
pillow=5.4.1=py37h34e0f95_0
pip=19.0.3=py37_0
python=3.7.0=h6e4f718_3
pytorch>=1.0.1
readline=7.0=h7b6447c_5
scikit-image=0.15.0=py37he1b5a44_0
scikit-learn=0.20.3=py37ha8026db_1
scipy=1.2.1=py37h09a28d5_1
setuptools=40.8.0=py37_0
torchvision>=0.2.2=py_3
werkzeug=0.15.2=py_0
wheel=0.33.1=py37_0
